Rename module: 'columns' -> 'columnar'
