+++ hypothesis_sqlalchemy/columnar/base.py
