+++ hypothesis_sqlalchemy/tables/base.py
@@ -7,7 +7,7 @@ from sqlalchemy.schema import (Column,
                                MetaData,
                                Table)
 
-from hypothesis_sqlalchemy import columns
+from hypothesis_sqlalchemy import columnar
 from hypothesis_sqlalchemy.hints import Strategy
 from hypothesis_sqlalchemy.utils import sql_identifiers
 
@@ -16,7 +16,7 @@ def factory(*,
             tables_names: Strategy[str] = sql_identifiers,
             metadatas: Strategy[MetaData],
             columns_lists: Strategy[List[Column]] =
-            columns.non_all_unique_lists_factory(),
+            columnar.non_all_unique_lists_factory(),
             extend_existing: Strategy[bool] = strategies.just(True)
             ) -> Strategy:
     def table_factory(draw: Callable[[Strategy], Any]) -> Table:
