+++ tests/tables_tests/records_tests/strategies.py
@@ -8,9 +8,9 @@ from hypothesis import strategies
 from sqlalchemy.schema import (Column,
                                Table)
 
-from hypothesis_sqlalchemy import (columns,
+from hypothesis_sqlalchemy import (columnar,
                                    tables)
-from hypothesis_sqlalchemy.columns import values
+from hypothesis_sqlalchemy.columnar import values
 from hypothesis_sqlalchemy.hints import Strategy
 from tests.strategies import (data,
                               max_sizes,
@@ -20,7 +20,7 @@ from tests.strategies import (data,
 data = data
 min_sizes = min_sizes
 max_sizes = max_sizes
-non_unique_columns = columns.non_primary_keys_factory(
+non_unique_columns = columnar.non_primary_keys_factory(
         are_unique=strategies.just(False))
 tables_without_unique_columns = tables.factory(
         metadatas=metadatas,
